Move config command out of obi main package
